adopts pr suggestion to push fork checking down into versioned impls and leave default case to valid
