+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineNewPayloadV1.java
@@ -69,9 +69,4 @@ public class EngineNewPayloadV1 extends AbstractEngineNewPayload {
       final ProtocolSpec protocolSpec) {
     return ValidationResult.valid();
   }
-
-  @Override
-  protected ValidationResult<RpcErrorType> validateForkSupported(final long blockTimestamp) {
-    return ValidationResult.valid();
-  }
 }
