+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineNewPayloadV2.java
@@ -57,9 +57,4 @@ public class EngineNewPayloadV2 extends AbstractEngineNewPayload {
       final ProtocolSpec protocolSpec) {
     return ValidationResult.valid();
   }
-
-  @Override
-  protected ValidationResult<RpcErrorType> validateForkSupported(final long blockTimestamp) {
-    return ValidationResult.valid();
-  }
 }
